Also check for executable /usr/sbin/slapcat in default cron-job
Fixes daily cron-job error mails if ldap-git-backup is installed, but
slapcat is not available, i.e. cases where it is just used manually
with data from remote hosts.

Reasoning:

Since ldap-git-backup has no configuration file or similar, a
different ldif-cmd must be configured on the commandline, i.e. in the
cron job itself, anyway.

So it is safe to also require /usr/sbin/slapcat to exist and be
executable in the default cron-job to avoid daily cron error mails,
because if someone uses a different command to get the LDIF data, the
cron-job has to be edited anyway, then also this additional constraint
can be removed or changed if necessary.