Add options --dn, --sort and --sort-prefix
If ldap-git-backup is used to backup the OpenLDAP configuration
stored in cn=config, the order of entries is critical for a restore
scenario. Therefore adding the possibility to have the entries' file
names prefixed with a zero-padded number will allow to concatenate
the files using their filename in the correct order. This may also
be useful if one needs to restore only some parts of the database
using ldapadd to a running slapd instance. It may also be useful to
have the DN of the entry in the filename to concatenate only parts
of the database backup.