Fix a bug where ARG_MAX is exceeded during git-add
When the argument length in the git-add operation, resulting from
all the filenames of all ldif files is greater than ARG_MAX, an
error occurs similar as when trying to list these files in the
backup directory after that using a wildcard `ls /backupdir/*`:
  -bash: /bin/ls: Argument list too long
On my testsystem a `getconf ARG_MAX` reports 2097152, but the
actual limit was about 1880000. Maybe this value has to be deter-
mined dynamically during runtime. See also:
https://unix.stackexchange.com/a/120842/60293